+++ Wave/Junction.xcodeproj/xcshareddata/xcschemes/Junction.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "39CBDD931D1D77B100CA2124"
-               BuildableName = "Wave.framework"
-               BlueprintName = "Wave"
-               ReferencedContainer = "container:Wave.xcodeproj">
+               BuildableName = "Junction.framework"
+               BlueprintName = "Junction"
+               ReferencedContainer = "container:Junction.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "39CBDD9D1D1D77B100CA2124"
-               BuildableName = "WaveTests.xctest"
-               BlueprintName = "WaveTests"
-               ReferencedContainer = "container:Wave.xcodeproj">
+               BuildableName = "Junction.xctest"
+               BlueprintName = "JunctionTests"
+               ReferencedContainer = "container:Junction.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -43,9 +43,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "39B377751D22C378009C4834"
-            BuildableName = "WaveExample.app"
-            BlueprintName = "WaveExample"
-            ReferencedContainer = "container:Wave.xcodeproj">
+            BuildableName = "Junction.app"
+            BlueprintName = "JunctionExample"
+            ReferencedContainer = "container:Junction.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -61,6 +61,15 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
+      <MacroExpansion>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "39CBDD931D1D77B100CA2124"
+            BuildableName = "Junction.framework"
+            BlueprintName = "Junction"
+            ReferencedContainer = "container:Junction.xcodeproj">
+         </BuildableReference>
+      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
@@ -74,9 +83,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "39CBDD931D1D77B100CA2124"
-            BuildableName = "Wave.framework"
-            BlueprintName = "Wave"
-            ReferencedContainer = "container:Wave.xcodeproj">
+            BuildableName = "Junction.framework"
+            BlueprintName = "Junction"
+            ReferencedContainer = "container:Junction.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
    </ProfileAction>
