+++ Wave/Junction.xcodeproj/xcuserdata/jimmymcdermott.xcuserdatad/xcschemes/JunctionExample.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "39B377751D22C378009C4834"
-               BuildableName = "WaveExample.app"
-               BlueprintName = "WaveExample"
-               ReferencedContainer = "container:Wave.xcodeproj">
+               BuildableName = "Junction.app"
+               BlueprintName = "JunctionExample"
+               ReferencedContainer = "container:Junction.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "39B377751D22C378009C4834"
-            BuildableName = "WaveExample.app"
-            BlueprintName = "WaveExample"
-            ReferencedContainer = "container:Wave.xcodeproj">
+            BuildableName = "Junction.app"
+            BlueprintName = "JunctionExample"
+            ReferencedContainer = "container:Junction.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -56,9 +56,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "39B377751D22C378009C4834"
-            BuildableName = "WaveExample.app"
-            BlueprintName = "WaveExample"
-            ReferencedContainer = "container:Wave.xcodeproj">
+            BuildableName = "Junction.app"
+            BlueprintName = "JunctionExample"
+            ReferencedContainer = "container:Junction.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -75,9 +75,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "39B377751D22C378009C4834"
-            BuildableName = "WaveExample.app"
-            BlueprintName = "WaveExample"
-            ReferencedContainer = "container:Wave.xcodeproj">
+            BuildableName = "Junction.app"
+            BlueprintName = "JunctionExample"
+            ReferencedContainer = "container:Junction.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
