+++ Wave/Junction.xcodeproj/xcuserdata/jimmymcdermott.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,12 +4,12 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>Wave.xcscheme_^#shared#^_</key>
+		<key>Junction.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
 		</dict>
-		<key>WaveExample.xcscheme</key>
+		<key>JunctionExample.xcscheme</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>1</integer>
