+++ Wave/Wave/Junction.h
@@ -1,6 +1,6 @@
 //
-//  Wave.h
-//  Wave
+//  Junction.h
+//  Junction
 //
 //  Created by Jimmy McDermott on 6/24/16.
 //  Copyright © 2016 BlueLine Labs. All rights reserved.
@@ -9,10 +9,10 @@
 #import <UIKit/UIKit.h>
 
 //! Project version number for Wave.
-FOUNDATION_EXPORT double WaveVersionNumber;
+FOUNDATION_EXPORT double JunctionVersionNumber;
 
 //! Project version string for Wave.
-FOUNDATION_EXPORT const unsigned char WaveVersionString[];
+FOUNDATION_EXPORT const unsigned char JunctionVersionString[];
 
 // In this header, you should import all the public headers of your framework using statements like #import <Wave/PublicHeader.h>
