+++ Wave/Wave/Section.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import UIKit
 
 public final class Section: SectionType {
     public var name: String = ""
