+++ Wave/Wave/Setting.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import UIKit
 
 public class Setting: SettingType, RowType {
