+++ Wave/Wave/StringSetting.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import UIKit
 
 public final class StringSetting: Setting {
     public var placeholder: String?
